+++ example/user/book/book_mongo.dart
@@ -13,8 +13,4 @@ part 'book_mongo.g.dart';
   'name': const EnDecode(alias: 'N'),
   'viewSerializer': ignore,
 })
-class BookMongoSerializer extends Serializer<Book> with _$BookMongoSerializer {
-  Book createModel() => new Book();
-
-  BookMongoSerializer();
-}
+class BookMongoSerializer extends Serializer<Book> with _$BookMongoSerializer {}