+++ lib/src/serializer/serializer.dart
@@ -53,14 +53,6 @@ abstract class Serializer<ModelType> {
     }
   }
 
-  /// Decodes model from [Map]
-  ModelType fromMap(Map map, {ModelType model}) {
-    if (model is! ModelType) {
-      model = createModel();
-    }
-    return model;
-  }
-
   /// Return the [Type] handle by this [Serializer]
   Type modelType() => ModelType;
 
@@ -72,7 +64,11 @@ abstract class Serializer<ModelType> {
 
   ////// To implement //////
 
-  ModelType createModel();
+  @Deprecated("release 1.0.0")
+  ModelType createModel() => null;
+
+  /// Decodes model from [Map]
+  ModelType fromMap(Map map, {ModelType model});
 
   /// Encodes model to [Map]
   Map toMap(ModelType model, {bool withType: false, String typeKey});