+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0-dev3
+
+- we now support final field deserialization
+- `createModel()` in `Serializer` class is now deprecated, it will be remove in the final release of `1.0.0`
+
 ## 1.0.0-dev2
 
 - simplify and optimize generated code